Refactoring to break out HTTP from WiFi
