+++ src/main.cpp
@@ -42,6 +42,7 @@
 #include <OneButton.h>
 #include <Wire.h>
 #include "meshwifi/meshwifi.h"
+#include "meshwifi/meshhttp.h"
 // #include <driver/rtc_io.h>
 
 #ifndef NO_ESP32
@@ -426,6 +427,15 @@ void loop()
     // TODO: This should go into a thread handled by FreeRTOS.
     handleWebResponse();
 #endif
-
+/*
+    const char *wifiName = radioConfig.preferences.wifi_ssid;
+    const char *wifiPsw = radioConfig.preferences.wifi_password;
+    Serial.print("-------------------");
+    Serial.print(wifiName);
+    Serial.print(" ");
+    Serial.println(wifiPsw);
+    Serial.println("+++++++++++++++++++");
+    Serial.println("");
+*/
     delay(msecstosleep);
 }
