+++ src/mesh/NodeDB.cpp
@@ -401,10 +401,11 @@ void NodeDB::updateFrom(const MeshPacket &mp)
                     powerFSM.trigger(EVENT_RECEIVED_TEXT_MSG);
                     notifyObservers(true); // Force an update whether or not our node counts have changed
 
+// This is going into the wifidev feature branch
 // Only update the WebUI if WiFi is enabled
-#if WiFi_MODE != 0
-                    notifyWebUI();
-#endif
+//#if WiFi_MODE != 0
+//  notifyWebUI();
+//#endif
                 }
             }
             break;
