+++ src/meshwifi/meshwifi.h
@@ -4,19 +4,9 @@
 #include <functional>
 #include <WiFi.h>
 
-void handleNotFound();
-
 void reconnectWiFi();
 
 void initWifi();
-
-void initWebServer();
-
-void handleWebResponse();
-
-void notifyWebUI();
-
-void handleJSONChatHistory();
 
 void WiFiEvent(WiFiEvent_t event);
